move retriever config to another package
