+++ cmd/relayproxy/controller/all_flags_test.go
@@ -15,6 +15,7 @@ import (
 	"github.com/thomaspoignant/go-feature-flag/cmd/relayproxy/controller"
 	"github.com/thomaspoignant/go-feature-flag/cmd/relayproxy/metric"
 	"github.com/thomaspoignant/go-feature-flag/cmd/relayproxy/service"
+	"github.com/thomaspoignant/go-feature-flag/cmdhelpers/retrieverconf"
 	"github.com/thomaspoignant/go-feature-flag/notifier"
 	"go.uber.org/zap"
 )
@@ -91,8 +92,8 @@ func Test_all_flag_Handler_DefaultMode(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			conf := config.Config{
 				CommonFlagSet: config.CommonFlagSet{
-					Retriever: &config.RetrieverConf{
-						Kind: config.FileRetriever,
+					Retriever: &retrieverconf.RetrieverConf{
+						Kind: retrieverconf.FileRetriever,
 						Path: tt.args.configFlagsLocation,
 					},
 					Exporter: &config.ExporterConf{
@@ -223,8 +224,8 @@ func Test_all_flag_Handler_FlagsetMode(t *testing.T) {
 					{
 						APIKeys: []string{"test-api-key"},
 						CommonFlagSet: config.CommonFlagSet{
-							Retriever: &config.RetrieverConf{
-								Kind: config.FileRetriever,
+							Retriever: &retrieverconf.RetrieverConf{
+								Kind: retrieverconf.FileRetriever,
 								Path: tt.args.configFlagsLocation,
 							},
 							Exporter: &config.ExporterConf{
