+++ cmd/relayproxy/controller/retriever_refresh_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/thomaspoignant/go-feature-flag/cmd/relayproxy/controller"
 	"github.com/thomaspoignant/go-feature-flag/cmd/relayproxy/metric"
 	"github.com/thomaspoignant/go-feature-flag/cmd/relayproxy/service"
+	"github.com/thomaspoignant/go-feature-flag/cmdhelpers/retrieverconf"
 	"github.com/thomaspoignant/go-feature-flag/notifier"
 	"go.uber.org/zap"
 )
@@ -35,8 +36,8 @@ func Test_retriever_refresh_Handler_valid(t *testing.T) {
 	// Create config for default mode
 	conf := config.Config{
 		CommonFlagSet: config.CommonFlagSet{
-			Retriever: &config.RetrieverConf{
-				Kind: config.FileRetriever,
+			Retriever: &retrieverconf.RetrieverConf{
+				Kind: retrieverconf.FileRetriever,
 				Path: "../../../testdata/flag-config.yaml",
 			},
 		},
