+++ cmd/relayproxy/service/gofeatureflag_client_test.go
@@ -5,7 +5,6 @@ import (
 	"crypto/tls"
 	"fmt"
 	"net"
-	"net/http"
 	"testing"
 	"time"
 
@@ -13,6 +12,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	ffclient "github.com/thomaspoignant/go-feature-flag"
 	"github.com/thomaspoignant/go-feature-flag/cmd/relayproxy/config"
+	"github.com/thomaspoignant/go-feature-flag/cmdhelpers/retrieverconf"
 	"github.com/thomaspoignant/go-feature-flag/exporter"
 	"github.com/thomaspoignant/go-feature-flag/exporter/azureexporter"
 	"github.com/thomaspoignant/go-feature-flag/exporter/fileexporter"
@@ -45,235 +45,16 @@ import (
 	"golang.org/x/net/proxy"
 )
 
-func Test_initRetriever(t *testing.T) {
-	tests := []struct {
-		name                   string
-		conf                   *config.RetrieverConf
-		want                   retriever.Retriever
-		wantErr                assert.ErrorAssertionFunc
-		wantType               retriever.Retriever
-		skipCompleteValidation bool
-	}{
-		{
-			name:    "Convert Github Retriever",
-			wantErr: assert.NoError,
-			conf: &config.RetrieverConf{
-				Kind:           "github",
-				RepositorySlug: "thomaspoignant/go-feature-flag",
-				Path:           "testdata/flag-config.yaml",
-				Timeout:        20,
-			},
-			want: &githubretriever.Retriever{
-				RepositorySlug: "thomaspoignant/go-feature-flag",
-				Branch:         "main",
-				FilePath:       "testdata/flag-config.yaml",
-				GithubToken:    "",
-				Timeout:        20 * time.Millisecond,
-			},
-			wantType: &githubretriever.Retriever{},
-		},
-		{
-			name:    "Convert Github Retriever with token",
-			wantErr: assert.NoError,
-			conf: &config.RetrieverConf{
-				Kind:           "github",
-				RepositorySlug: "thomaspoignant/go-feature-flag",
-				Path:           "testdata/flag-config.yaml",
-				Timeout:        20,
-				AuthToken:      "xxx",
-			},
-			want: &githubretriever.Retriever{
-				RepositorySlug: "thomaspoignant/go-feature-flag",
-				Branch:         "main",
-				FilePath:       "testdata/flag-config.yaml",
-				GithubToken:    "xxx",
-				Timeout:        20 * time.Millisecond,
-			},
-			wantType: &githubretriever.Retriever{},
-		},
-		{
-			name:    "Convert Github Retriever with deprecated token",
-			wantErr: assert.NoError,
-			conf: &config.RetrieverConf{
-				Kind:           "github",
-				RepositorySlug: "thomaspoignant/go-feature-flag",
-				Path:           "testdata/flag-config.yaml",
-				Timeout:        20,
-				GithubToken:    "xxx",
-			},
-			want: &githubretriever.Retriever{
-				RepositorySlug: "thomaspoignant/go-feature-flag",
-				Branch:         "main",
-				FilePath:       "testdata/flag-config.yaml",
-				GithubToken:    "xxx",
-				Timeout:        20 * time.Millisecond,
-			},
-			wantType: &githubretriever.Retriever{},
-		},
-		{
-			name:    "Convert Gitlab Retriever",
-			wantErr: assert.NoError,
-			conf: &config.RetrieverConf{
-				Kind:           "gitlab",
-				BaseURL:        "http://localhost",
-				Path:           "flag-config.yaml",
-				RepositorySlug: "thomaspoignant/go-feature-flag",
-				Timeout:        20,
-			},
-			want: &gitlabretriever.Retriever{
-				BaseURL:        "http://localhost",
-				Branch:         "main",
-				FilePath:       "flag-config.yaml",
-				RepositorySlug: "thomaspoignant/go-feature-flag",
-				GitlabToken:    "",
-				Timeout:        20 * time.Millisecond,
-			},
-			wantType: &gitlabretriever.Retriever{},
-		},
-		{
-			name:    "Convert File Retriever",
-			wantErr: assert.NoError,
-			conf: &config.RetrieverConf{
-				Kind: "file",
-				Path: "testdata/flag-config.yaml",
-			},
-			want:     &fileretriever.Retriever{Path: "testdata/flag-config.yaml"},
-			wantType: &fileretriever.Retriever{},
-		},
-		{
-			name:    "Convert S3 Retriever",
-			wantErr: assert.NoError,
-			conf: &config.RetrieverConf{
-				Kind:   "s3",
-				Bucket: "my-bucket-name",
-				Item:   "testdata/flag-config.yaml",
-			},
-			want: &s3retrieverv2.Retriever{
-				Bucket: "my-bucket-name",
-				Item:   "testdata/flag-config.yaml",
-			},
-			wantType:               &s3retrieverv2.Retriever{},
-			skipCompleteValidation: true,
-		},
-		{
-			name:    "Convert HTTP Retriever",
-			wantErr: assert.NoError,
-			conf: &config.RetrieverConf{
-				Kind: "http",
-				URL:  "https://gofeatureflag.org/my-flag-test.yaml",
-			},
-			want: &httpretriever.Retriever{
-				URL:     "https://gofeatureflag.org/my-flag-test.yaml",
-				Method:  http.MethodGet,
-				Body:    "",
-				Header:  nil,
-				Timeout: 10000000000,
-			},
-			wantType: &httpretriever.Retriever{},
-		},
-		{
-			name:    "Convert Google storage Retriever",
-			wantErr: assert.NoError,
-			conf: &config.RetrieverConf{
-				Kind:   "googleStorage",
-				Bucket: "my-bucket-name",
-				Object: "testdata/flag-config.yaml",
-			},
-			want: &gcstorageretriever.Retriever{
-				Bucket:  "my-bucket-name",
-				Object:  "testdata/flag-config.yaml",
-				Options: nil,
-			},
-			wantType: &gcstorageretriever.Retriever{},
-		},
-		{
-			name:    "Convert unknown Retriever",
-			wantErr: assert.Error,
-			conf: &config.RetrieverConf{
-				Kind: "unknown",
-			},
-		},
-		{
-			name:    "Convert Bitbucket Retriever default branch",
-			wantErr: assert.NoError,
-			conf: &config.RetrieverConf{
-				Kind:           "bitbucket",
-				RepositorySlug: "gofeatureflag/config-repo",
-				Path:           "flags/config.goff.yaml",
-				AuthToken:      "XXX_BITBUCKET_TOKEN",
-				BaseURL:        "https://api.bitbucket.goff.org",
-			},
-			want: &bitbucketretriever.Retriever{
-				RepositorySlug: "gofeatureflag/config-repo",
-				Branch:         "main",
-				FilePath:       "flags/config.goff.yaml",
-				BitBucketToken: "XXX_BITBUCKET_TOKEN",
-				BaseURL:        "https://api.bitbucket.goff.org",
-				Timeout:        10000000000,
-			},
-			wantType: &bitbucketretriever.Retriever{},
-		},
-		{
-			name:    "Convert Bitbucket Retriever branch specified",
-			wantErr: assert.NoError,
-			conf: &config.RetrieverConf{
-				Kind:           "bitbucket",
-				Branch:         "develop",
-				RepositorySlug: "gofeatureflag/config-repo",
-				Path:           "flags/config.goff.yaml",
-				AuthToken:      "XXX_BITBUCKET_TOKEN",
-				BaseURL:        "https://api.bitbucket.goff.org",
-			},
-			want: &bitbucketretriever.Retriever{
-				RepositorySlug: "gofeatureflag/config-repo",
-				Branch:         "develop",
-				FilePath:       "flags/config.goff.yaml",
-				BitBucketToken: "XXX_BITBUCKET_TOKEN",
-				BaseURL:        "https://api.bitbucket.goff.org",
-				Timeout:        10000000000,
-			},
-			wantType: &bitbucketretriever.Retriever{},
-		},
-		{
-			name:    "Convert Postgres Retriever",
-			wantErr: assert.NoError,
-			conf: &config.RetrieverConf{
-				Kind:    "postgresql",
-				URI:     "postgresql://user:password@localhost:5432/database",
-				Table:   "flags",
-				Columns: map[string]string{"flagset": "settings"},
-			},
-			want: &postgresqlretriever.Retriever{
-				URI:     "postgresql://user:password@localhost:5432/database",
-				Table:   "flags",
-				Columns: map[string]string{"flagset": "settings"},
-			},
-			wantType: &postgresqlretriever.Retriever{},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := initRetriever(tt.conf)
-			tt.wantErr(t, err)
-			if err == nil {
-				assert.IsType(t, tt.wantType, got)
-				if !tt.skipCompleteValidation {
-					assert.Equal(t, tt.want, got)
-				}
-			}
-		})
-	}
-}
 func Test_initRetrievers(t *testing.T) {
 	tests := []struct {
 		name       string
-		retrievers *[]config.RetrieverConf
+		retrievers *[]retrieverconf.RetrieverConf
 		wantErr    assert.ErrorAssertionFunc
 	}{
 		{
 			name:    "both retriever and retrievers",
 			wantErr: assert.NoError,
-			retrievers: &[]config.RetrieverConf{
+			retrievers: &[]retrieverconf.RetrieverConf{
 				{
 					Kind:           "bitbucket",
 					Branch:         "develop",
@@ -295,7 +76,7 @@ func Test_initRetrievers(t *testing.T) {
 		{
 			name:    "should error with invalid retriever",
 			wantErr: assert.Error,
-			retrievers: &[]config.RetrieverConf{
+			retrievers: &[]retrieverconf.RetrieverConf{
 				{
 					Kind:           "bitbucket",
 					Branch:         "develop",
@@ -312,7 +93,7 @@ func Test_initRetrievers(t *testing.T) {
 		{
 			name:    "should error with invalid retriever",
 			wantErr: assert.Error,
-			retrievers: &[]config.RetrieverConf{
+			retrievers: &[]retrieverconf.RetrieverConf{
 				{
 					Kind: "unknown",
 				},
@@ -321,7 +102,7 @@ func Test_initRetrievers(t *testing.T) {
 		{
 			name:    "only retrievers",
 			wantErr: assert.NoError,
-			retrievers: &[]config.RetrieverConf{
+			retrievers: &[]retrieverconf.RetrieverConf{
 				{
 					Kind:           "bitbucket",
 					Branch:         "develop",
