+++ cmdhelpers/retrieverconf/retriever_conf.go
@@ -1,4 +1,4 @@
-package config
+package retrieverconf
 
 import (
 	"fmt"
